# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlazeOS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Trade {
  id: number;
  symbol: string;
  direction: string;
  entryPrice: number;
  exitPrice: number;
  riskPercent: number;
  lotSize: number;
  setupType: string;
  session: string;
  result: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  screenshotBefore?: string | null;
  /** @nullable */
  screenshotAfter?: string | null;
  /** @nullable */
  pnl?: number | null;
  createdAt: string;
}

export interface DashboardSummary {
  totalTrades: number;
  winRate: number;
  totalPnl: number;
  openClients: number;
  totalNotes: number;
  upcomingEvents: number;
  monthlyRevenue: number;
  recentTrades?: Trade[];
}

export interface DailyBias {
  id: number;
  direction: string;
  /** @nullable */
  notes: string | null;
  updatedAt: string;
}

export interface DailyBiasInput {
  direction: string;
  notes?: string;
}

export interface TradeInput {
  symbol: string;
  direction: string;
  entryPrice: number;
  exitPrice: number;
  riskPercent: number;
  lotSize: number;
  setupType: string;
  session: string;
  result: string;
  notes?: string;
  screenshotBefore?: string;
  screenshotAfter?: string;
  pnl?: number;
}

export interface TradeUpdate {
  symbol?: string;
  direction?: string;
  entryPrice?: number;
  exitPrice?: number;
  riskPercent?: number;
  lotSize?: number;
  setupType?: string;
  session?: string;
  result?: string;
  notes?: string;
  screenshotBefore?: string;
  screenshotAfter?: string;
  pnl?: number;
}

export interface TradeStatBySymbol {
  symbol: string;
  count: number;
  winRate: number;
  pnl: number;
}

export interface TradeStatBySetup {
  setupType: string;
  count: number;
  winRate: number;
}

export interface TradeStats {
  totalTrades: number;
  wins: number;
  losses: number;
  winRate: number;
  totalPnl: number;
  avgRisk: number;
  /** @nullable */
  bestTrade: number | null;
  /** @nullable */
  worstTrade: number | null;
  bySymbol?: TradeStatBySymbol[];
  bySetup?: TradeStatBySetup[];
}

export interface Client {
  id: number;
  name: string;
  businessType: string;
  /** @nullable */
  contactInfo?: string | null;
  status: string;
  paymentStatus: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  projectValue?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export interface ClientInput {
  name: string;
  businessType: string;
  contactInfo?: string;
  status: string;
  paymentStatus: string;
  notes?: string;
  projectValue?: number;
}

export interface ClientUpdate {
  name?: string;
  businessType?: string;
  contactInfo?: string;
  status?: string;
  paymentStatus?: string;
  notes?: string;
  projectValue?: number;
}

export interface PipelineStage {
  stage: string;
  count: number;
  value: number;
  clients?: Client[];
}

export interface ClientPipeline {
  stages: PipelineStage[];
  totalValue?: number;
}

export interface Note {
  id: number;
  title: string;
  category: string;
  content: string;
  /** @nullable */
  tags?: string | null;
  isPinned?: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  title: string;
  category: string;
  content: string;
  tags?: string;
  isPinned?: boolean;
}

export interface NoteUpdate {
  title?: string;
  category?: string;
  content?: string;
  tags?: string;
  isPinned?: boolean;
}

export interface CalendarEvent {
  id: number;
  eventTime: string;
  currency: string;
  eventName: string;
  impact: string;
  /** @nullable */
  forecast?: string | null;
  /** @nullable */
  previous?: string | null;
  /** @nullable */
  actual?: string | null;
  createdAt: string;
}

export interface CalendarEventInput {
  eventTime: string;
  currency: string;
  eventName: string;
  impact: string;
  forecast?: string;
  previous?: string;
  actual?: string;
}

export interface CalendarEventUpdate {
  eventTime?: string;
  currency?: string;
  eventName?: string;
  impact?: string;
  forecast?: string;
  previous?: string;
  actual?: string;
}

export interface Finance {
  id: number;
  type: string;
  category: string;
  amount: number;
  description: string;
  month: string;
  createdAt: string;
}

export interface FinanceInput {
  type: string;
  category: string;
  amount: number;
  description: string;
  month: string;
}

export interface FinanceUpdate {
  type?: string;
  category?: string;
  amount?: number;
  description?: string;
  month?: string;
}

export interface MonthlyFinance {
  month: string;
  revenue: number;
  expenses: number;
  profit: number;
}

export interface FinanceByCategory {
  category: string;
  total: number;
  type: string;
}

export interface FinanceSummary {
  totalRevenue: number;
  totalExpenses: number;
  netProfit: number;
  monthlyBreakdown: MonthlyFinance[];
  byCategory?: FinanceByCategory[];
}

export type ListTradesParams = {
symbol?: string;
result?: string;
session?: string;
};

export type ListClientsParams = {
status?: string;
};

export type ListNotesParams = {
category?: string;
q?: string;
};

export type ListCalendarEventsParams = {
impact?: string;
currency?: string;
};

export type ListFinancesParams = {
type?: string;
month?: string;
};