# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BlazeOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all CRM clients
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
...
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a single client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── KNOWLEDGE VAULT NOTES ────────────────────────────────
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all knowledge vault notes
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a knowledge note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a single note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ECONOMIC CALENDAR ────────────────────────────────────
  /calendar:
    get:
      operationId: listCalendarEvents
      tags: [calendar]
      summary: List economic calendar events
      parameters:
        - in: query
          name: impact
          schema:
            type: string
        - in: query
          name: currency
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createCalendarEvent
      tags: [calendar]
      summary: Create an economic calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /calendar/{id}:
    patch:
      operationId: updateCalendarEvent
      tags: [calendar]
      summary: Update a calendar event (e.g. add actual value)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteCalendarEvent
      tags: [calendar]
      summary: Delete a calendar event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FINANCES ─────────────────────────────────────────────
  /finances:
    get:
      operationId: listFinances
      tags: [finances]
      summary: List all finance entries
      parameters:
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: month
          schema:
            type: string
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finance"
    post:
      operationId: createFinance
      tags: [finances]
      summary: Create a finance entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finance"
...
  /finances/{id}:
    patch:
      operationId: updateFinance
      tags: [finances]
      summary: Update a finance entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finance"
    delete:
      operationId: deleteFinance
      tags: [finances]
      summary: Delete a finance entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, businessType, status, paymentStatus, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        businessType: { type: string }
        contactInfo: { type: ["string", "null"] }
        status: { type: string }
        paymentStatus: { type: string }
        notes: { type: ["string", "null"] }
        projectValue: { type: ["number", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ClientInput:
      type: object
      required: [name, businessType, status, paymentStatus]
      properties:
        name: { type: string }
        businessType: { type: string }
        contactInfo: { type: string }
        status: { type: string }
        paymentStatus: { type: string }
        notes: { type: string }
        projectValue: { type: number }
    ClientUpdate:
      type: object
      properties:
        name: { type: string }
        businessType: { type: string }
        contactInfo: { type: string }
        status: { type: string }
        paymentStatus: { type: string }
        notes: { type: string }
        projectValue: { type: number }
...
    Note:
      type: object
      required: [id, title, category, content, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        category: { type: string }
        content: { type: string }
        tags: { type: ["string", "null"] }
        isPinned: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NoteInput:
      type: object
      required: [title, category, content]
      properties:
        title: { type: string }
        category: { type: string }
        content: { type: string }
        tags: { type: string }
        isPinned: { type: boolean }
    NoteUpdate:
      type: object
      properties:
        title: { type: string }
        category: { type: string }
        content: { type: string }
        tags: { type: string }
        isPinned: { type: boolean }

    # ─── CALENDAR ──────────────────────────────────────────
    CalendarEvent:
      type: object
      required: [id, eventTime, currency, eventName, impact, createdAt]
      properties:
        id: { type: integer }
        eventTime: { type: string }
        currency: { type: string }
        eventName: { type: string }
        impact: { type: string }
        forecast: { type: ["string", "null"] }
        previous: { type: ["string", "null"] }
        actual: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CalendarEventInput:
      type: object
      required: [eventTime, currency, eventName, impact]
      properties:
        eventTime: { type: string }
        currency: { type: string }
        eventName: { type: string }
        impact: { type: string }
        forecast: { type: string }
        previous: { type: string }
        actual: { type: string }
    CalendarEventUpdate:
      type: object
      properties:
        eventTime: { type: string }
        currency: { type: string }
        eventName: { type: string }
        impact: { type: string }
        forecast: { type: string }
        previous: { type: string }
        actual: { type: string }

    # ─── FINANCES ──────────────────────────────────────────
    Finance:
      type: object
      required: [id, type, category, amount, description, month, createdAt]
      properties:
        id: { type: integer }
        type: { type: string }
        category: { type: string }
        amount: { type: number }
        description: { type: string }
        month: { type: string }
        createdAt: { type: string, format: date-time }
    FinanceInput:
      type: object
      required: [type, category, amount, description, month]
      properties:
        type: { type: string }
        category: { type: string }
        amount: { type: number }
        description: { type: string }
        month: { type: string }
    FinanceUpdate:
      type: object
      properties:
        type: { type: string }
        category: { type: string }
        amount: { type: number }
        description: { type: string }
        month: { type: string }
...