# lib/api-zod/src/generated/types/financeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlazeOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FinanceInput {
  type: string;
  category: string;
  amount: number;
  description: string;
  month: string;
}