# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlazeOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NoteUpdate {
  title?: string;
  category?: string;
  content?: string;
  tags?: string;
  isPinned?: boolean;
}